% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_data_type_to_simple.R
\name{convert_data_type_to_simple}
\alias{convert_data_type_to_simple}
\title{Simplify data types for humans}
\usage{
convert_data_type_to_simple(.data)
}
\arguments{
\item{.data}{dataframe}
}
\value{
character vector of length ncol(.data)
}
\description{
Categorizes R data types into 'Continuous', 'Categorical', or 'Binary'
}
\examples{
n_row <- 10
my_character = sample(c('one', 'two', 'three'), size = n_row, replace = TRUE)
my_logical = sample(c(TRUE, FALSE), size = n_row, replace = TRUE)
my_numeric = rnorm(n_row)
X <- data.frame(my_character = my_character, my_logical = my_logical, my_numeric = my_numeric)
convert_data_type_to_simple(X)
}
\seealso{
\code{\link{convert_data_type_to_complex}}
}
\author{
Joe Marlo
}
